+++ test/jwt.hs.tests.js
@@ -1,11 +1,35 @@
-var jwt = require('../index');
+const jwt = require('../index');
 
-var expect = require('chai').expect;
-var assert = require('chai').assert;
+const jws = require('jws');
+const expect = require('chai').expect;
+const assert = require('chai').assert;
+const { generateKeyPairSync } = require('crypto')
 
 describe('HS256', function() {
 
-  describe('when signing a token', function() {
+  describe("when signing using HS256", function () {
+    it('should throw if the secret is an asymmetric key', function () {
+      const { privateKey } = generateKeyPairSync('rsa', { modulusLength: 2048 });
+
+      expect(function () {
+        jwt.sign({ foo: 'bar' }, privateKey, { algorithm: 'HS256' })
+      }).to.throw(Error, 'must be a symmetric key')
+    })
+
+    it('should throw if the payload is undefined', function () {
+      expect(function () {
+        jwt.sign(undefined, "secret", { algorithm: 'HS256' })
+      }).to.throw(Error, 'payload is required')
+    })
+
+    it('should throw if options is not a plain object', function () {
+      expect(function () {
+        jwt.sign({ foo: 'bar' }, "secret", ['HS256'])
+      }).to.throw(Error, 'Expected "options" to be a plain object')
+    })
+  })
+
+  describe('with a token signed using HS256', function() {
     var secret = 'shhhhhh';
 
     var token = jwt.sign({ foo: 'bar' }, secret, { algorithm: 'HS256' });
@@ -42,19 +66,21 @@ describe('HS256', function() {
     });
 
     it('should throw with secret and token not signed', function(done) {
-      var signed = jwt.sign({ foo: 'bar' }, secret, { algorithm: 'none' });
-      var unsigned = signed.split('.')[0] + '.' + signed.split('.')[1] + '.';
-      jwt.verify(unsigned, 'secret', function(err, decoded) {
+      const header = { alg: 'none' };
+      const payload = { foo: 'bar' };
+      const token = jws.sign({ header, payload, secret: 'secret', encoding: 'utf8' });
+      jwt.verify(token, 'secret', function(err, decoded) {
         assert.isUndefined(decoded);
         assert.isNotNull(err);
         done();
       });
     });
 
-    it('should work with falsy secret and token not signed', function(done) {
-      var signed = jwt.sign({ foo: 'bar' }, null, { algorithm: 'none' });
-      var unsigned = signed.split('.')[0] + '.' + signed.split('.')[1] + '.';
-      jwt.verify(unsigned, 'secret', function(err, decoded) {
+    it('should throw with falsy secret and token not signed', function(done) {
+      const header = { alg: 'none' };
+      const payload = { foo: 'bar' };
+      const token = jws.sign({ header, payload, secret: null, encoding: 'utf8' });
+      jwt.verify(token, 'secret', function(err, decoded) {
         assert.isUndefined(decoded);
         assert.isNotNull(err);
         done();